Allow deploys to servers while initialization is occurring
We use `cap-ec2` in conjunction with `ec2` autoscaling. As mentioned in the `README`, it takes a while for just-launched instances to make it to the `OK` status. This change makes it so that deploys can happen while they are still in an `initializing` state.

The problem with turning off `filter_by_status_ok?` completely is that trying to deploy to instances while they are `terminating` can cause a Capistrano deploy to hang indefinitely while it waits for a ssh connection that will never come. When triggering a deploy due to an auto-scaling action, there is almost always a terminating instance and an initializing instance at the same time.